requirements.txt.txt
openai
requests
time
sounddevice
pydub
Pillow
numpy
wave
moviepy
stability_sdk